added confirmation to comment delete link

+++ app/helpers/comments_helper.rb
@@ -5,7 +5,7 @@ module CommentsHelper
 
     if current_user == comment.user
       links.push link_to("Edit", discussion_comment_path(comment.discussion, comment))
-      links.push link_to("Delete", discussion_comment_path(comment.discussion, comment), method: :delete)
+      links.push link_to("Delete", discussion_comment_path(comment.discussion, comment), method: :delete, data: { confirm: "Are you sure?" })
     end
 
     action_links links
